+++ tests/Auth.Application.UT/Roles/Commands/DisabledRoleTest.cs
@@ -1,5 +1,6 @@
-﻿using Auth.Application.Exceptions;
-using Auth.Application.Roles.Commands.Create;
+﻿using Auth.Application.Contracts;
+using Auth.Application.Exceptions;
+using Auth.Application.Roles.Commands.Disabled;
 using Auth.Application.UT.Common;
 using FluentAssertions;
 using FluentValidation;
@@ -9,22 +10,21 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Xunit;
-
-namespace Auth.Application.UT.Roles.Commans
+namespace Auth.Application.UT.Roles.Commands
 {
     [ExcludeFromCodeCoverage]
-    public class CreateRoleTest : BaseTest
+    public class DisabledRoleTest : BaseTest
     {
         [Theory]
         [InlineData(null)]
         [InlineData("")]
         [InlineData("Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam sed tincidunt magna, ac consequat mauris. Praesent turpis augue, laoreet sed justo ut, efficitur euismod tortor. Ut laoreet nec ex nunc asdsdasdas das asdasdasdasdas")]
-        public async Task When_CreateRole_InputInValid_ThrowValidationException(string roleName)
+        public async Task When_DisabledRole_InputInValid_ThrowValidationException(string roleName)
         {
             var mediator = ServiceProvider.GetService<IMediator>();
-            Func<Task<string>> act = async () =>
-            {
-                var response = await mediator.Send(new CreateRoleCommand()
+            Func<Task<Unit>> act = async () =>
+            {                
+                var response = await mediator.Send(new DisabledRoleCommand()
                 {
                     Name = roleName
                 });
@@ -35,15 +35,18 @@ namespace Auth.Application.UT.Roles.Commans
         }
 
         [Theory]
-        [InlineData(Constants.RoleAdmin + "_test")]
-        [InlineData(Constants.RoleGuest + "_test")]
-        public async Task When_CreateRole_InputIsValid_Return(string roleName)
+        [InlineData(Constants.RoleAdmin)]
+        [InlineData(Constants.RoleGuest)]
+        public async Task When_DisabledRole_InputIsValid_Return(string roleName)
         {
             using var scope = ServiceScopeProvider.CreateScope();
             var sp = scope.ServiceProvider;
+
             var mediator = sp.GetService<IMediator>();
+            var rolemanager = sp.GetService<IAppDbContext>();
+           
             //Act
-            var response = await mediator.Send(new CreateRoleCommand()
+            var response = await mediator.Send(new DisabledRoleCommand()
             {
                 Name = roleName
             });
@@ -53,25 +56,24 @@ namespace Auth.Application.UT.Roles.Commans
         }
 
         [Theory]
-        [InlineData(Constants.RoleAdmin)]
-        [InlineData(Constants.RoleGuest)]
-        public async Task When_CreateRole_InputIsValid_ThrowExistsException(string roleName)
+        [InlineData(Constants.RoleAdmin + "1")]
+        [InlineData(Constants.RoleGuest + "guest1")]
+        public async Task When_DisabledRole_InputIsValid_ThrowNotFoundException(string roleName)
         {
             using var scope = ServiceScopeProvider.CreateScope();
             var sp = scope.ServiceProvider;
-
             var mediator = sp.GetService<IMediator>();
             //Act
-            Func<Task<string>> act = async () =>
-            {
-                var response = await mediator.Send(new CreateRoleCommand()
+            Func<Task<Unit>> act = async () =>
+            {                
+                var response = await mediator.Send(new DisabledRoleCommand()
                 {
                     Name = roleName
                 });
                 return response;
             };
             //Assert
-            act.Should().Throw<ExistsException>();
+            act.Should().Throw<NotFoundException>();
 
         }
     }
