+++ tests/Auth.Application.UT/Roles/Commands/EnabledRoleTest.cs
@@ -1,6 +1,6 @@
 ﻿using Auth.Application.Contracts;
 using Auth.Application.Exceptions;
-using Auth.Application.Roles.Commands.Delete;
+using Auth.Application.Roles.Commands.Enabled;
 using Auth.Application.UT.Common;
 using FluentAssertions;
 using FluentValidation;
@@ -10,21 +10,21 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Xunit;
-namespace Auth.Application.UT.Roles.Commans
+namespace Auth.Application.UT.Roles.Commands
 {
     [ExcludeFromCodeCoverage]
-    public class DisabledRoleTest : BaseTest
+    public class EnabledRoleTest : BaseTest
     {
         [Theory]
         [InlineData(null)]
         [InlineData("")]
         [InlineData("Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam sed tincidunt magna, ac consequat mauris. Praesent turpis augue, laoreet sed justo ut, efficitur euismod tortor. Ut laoreet nec ex nunc asdsdasdas das asdasdasdasdas")]
-        public async Task When_DisabledRole_InputInValid_ThrowValidationException(string roleName)
+        public async Task When_EnabledRole_InputInValid_ThrowValidationException(string roleName)
         {
             var mediator = ServiceProvider.GetService<IMediator>();
             Func<Task<Unit>> act = async () =>
-            {                
-                var response = await mediator.Send(new DisabledApplicationCommand()
+            {
+                var response = await mediator.Send(new EnabledRoleCommand()
                 {
                     Name = roleName
                 });
@@ -37,16 +37,16 @@ namespace Auth.Application.UT.Roles.Commans
         [Theory]
         [InlineData(Constants.RoleAdmin)]
         [InlineData(Constants.RoleGuest)]
-        public async Task When_DisabledRole_InputIsValid_Return(string roleName)
+        public async Task When_EnabledRole_InputIsValid_Return(string roleName)
         {
             using var scope = ServiceScopeProvider.CreateScope();
             var sp = scope.ServiceProvider;
 
             var mediator = sp.GetService<IMediator>();
             var rolemanager = sp.GetService<IAppDbContext>();
-           
+
             //Act
-            var response = await mediator.Send(new DisabledApplicationCommand()
+            var response = await mediator.Send(new EnabledRoleCommand()
             {
                 Name = roleName
             });
@@ -58,15 +58,15 @@ namespace Auth.Application.UT.Roles.Commans
         [Theory]
         [InlineData(Constants.RoleAdmin + "1")]
         [InlineData(Constants.RoleGuest + "guest1")]
-        public async Task When_DisabledRole_InputIsValid_ThrowNotFoundException(string roleName)
+        public async Task When_EnabledRole_InputIsValid_ThrowNotFoundException(string roleName)
         {
             using var scope = ServiceScopeProvider.CreateScope();
             var sp = scope.ServiceProvider;
             var mediator = sp.GetService<IMediator>();
             //Act
             Func<Task<Unit>> act = async () =>
-            {                
-                var response = await mediator.Send(new DisabledApplicationCommand()
+            {
+                var response = await mediator.Send(new EnabledRoleCommand()
                 {
                     Name = roleName
                 });
